Moved /pages route out of app.js
