+++ app.js
@@ -2,7 +2,7 @@
 var bodyParser = require('body-parser'),
     cookieParser = require('cookie-parser'),
     dateFunctions = require('./functions/date-functions.js'),
-    bandFunctions = require('./functions/fetchBandInfo.js'),
+    // bandFunctions = require('./functions/fetchBandInfo.js'),
     express = require('express'),
     favicon = require('serve-favicon'),
     hbs = require('hbs'),
@@ -82,104 +82,6 @@ app.use(require('express-session')({
 }));
 
 app.set('port', (process.env.PORT || 5000));
-
-// Route that's called to access any public itinerary
-app.get('/pages/*', function(req, res) {
-
-    try{
-
-    // Find user with name "*" in Users database. "*" is accessed with req.params[0]
-    User.find({name: req.params[0]}).exec(function(err, docs){
-
-      if(err){
-          return res.send({errors: "Error connecting to our database"});
-          } else{
-            try{
-
-                // Check to see if current page matches the user's logged in session
-                var requestSession = req.session.id;
-                var userSession = docs[0].uid;
-                var displayName = docs[0].displayName;
-                var public = docs[0].public;
-                var link = "http://sxdiscover.co/pages/"+req.params[0];
-
-                console.log(link);
-
-                var fbShareMessage = "My SXSW 🎉"
-                var displayMessage = "Viewing someone else's itinerary"
-
-                if(displayName != null){
-                  fbShareMessage = displayName+"'s SXSW 🎉";
-                  displayMessage = "Viewing "+displayName+"'s itinerary";
-                }
-
-                 // Check to see if page is set to public or private, or if user visiting their own page
-
-                var results = bandFunctions.fetchBandInfo(docs[0].rawBands, function(result){
-
-                  var r = JSON.stringify(result);
-                  console.log("JSON result is");
-                  console.log(r);
-
-                  if(public==true){
-                    if(requestSession==userSession){
-                      // If user is visiting their own page, show the delete button
-                      res.render('userPage', {displayMessage:displayMessage, bands:r, canDelete:'y', public:'y', status:"public", link:link, fbShareMessage:fbShareMessage});
-                    } else {
-                      res.render('userPage', {displayMessage:displayMessage, bands:r, canDelete:'n', public:'y', status:"public", link:link, fbShareMessage:fbShareMessage});
-                    }
-                  } else if(requestSession==userSession) {
-                    res.render('userPage', {displayMessage:displayMessage, bands:r, canDelete:'y', public:'n', status:"private"});
-                  } else {
-                    return res.render('privateUser');
-                  }
-                });
-            } catch (e){
-              console.log("Error is " + e);
-              return res.send({errors: "User page doesn't exist"});
-            }
-          }
-    });
-  } catch (e){
-    res.send(404);
-  }
-});
-
-// app.get('/private', function(req, res){
-
-//   // By default, set public to true
-//   var status = true
-//   var statusStr = "public";
-
-//   // Check to see user's public status
-//   User.find({"uid": req.session.id}).exec(function(err, docs){
-//     if(err){
-//       return res.send({errors:"Error finding profile. If issue persists, please message us on Facebook."})
-//     }
-
-//     status = docs[0].public;
-
-//     // Toggle status
-//     if (status == true){
-//       status = false;
-//       statusStr = "private"
-//     } else {
-//       status = true;
-//       statusStr = "public"
-//     }
-
-//     console.log(req.session.id);
-
-//     // Update status
-//     User.update({ "uid" : req.session.id }, {$set: {"public" : status}}, function(err){
-//     if(err){
-//       console.log("Error: Failed to make user private");
-//       return res.send({errors: "Failed to make itinerary private. If issue persists, please message us on Facebook."});
-//     }
-//       return res.send("Page is now "+statusStr+".");
-//     });
-//   });
-// });
 
 // uncomment after placing your favicon in /public
 app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
